Add fix for centos in 4.2
In order for machine config to work on centos
and 4.2, we need to update the version of logger.
It is because machine config uses a --journal
flag that is not brought with the default version,
so upgrade it. The only repos that are bringing
this rpm is koji, and need to be installed as
individual rpms.
This is temporary until the fix to disable journal
lands properly.