Filter images_and_binaries by version
Starting with 4.5.6, the OpenShift mirror began including image files
with and without version numbers in the sha256sum listings*.  This
caused two matches to occur in the images_and_binaries script, producing
data that could not be processed correctly by further scripts such
as gen_matchbox.

To correct this, filter the filenames by version number so that new
and old sha256sum files produce the same behavior.

* See these files for examples:
  https://mirror.openshift.com/pub/openshift-v4/dependencies/rhcos/4.5/4.5.6/sha256sum.txt
  https://mirror.openshift.com/pub/openshift-v4/dependencies/rhcos/4.5/4.5.2/sha256sum.txt

Signed-off-by: James E. Blair <jeblair@redhat.com>